fixed bugs

+++ src/installies/models/voting.py
@@ -21,8 +21,7 @@ class VoteJunction(BaseModel):
 
         Also updates the score field.
         """
-        if self.has_user_voted(user):
-            self.remove_vote(user)
+        self.remove_vote(user)
         
         Vote.create(
             user=user,
@@ -40,8 +39,7 @@ class VoteJunction(BaseModel):
 
         Also updates the score field.
         """
-        if self.has_user_voted(user):
-            self.remove_vote(user)
+        self.remove_vote(user)
         
         Vote.create(
             user=user,
@@ -59,25 +57,26 @@ class VoteJunction(BaseModel):
 
         Also updates the score field.
         """
+        if has_user_voted(user) is False:
+            return
+        
         vote = self.votes.where(Vote.user == user).get().delete_instance()
 
         # updates the vote count
         self.score = self.get_score()
         self.save()
 
-        return vote
-
     def has_user_upvoted(self, user: User):
         """Checks if a given user has upvoted."""
-        return self.votes.where(Vote.user == user & Vote.is_upvote).exists()
+        return self.votes.where(Vote.user == user).where(Vote.is_upvote).exists()
 
     def has_user_downvoted(self, user: User):
         """Checks if a given user has downvoted."""
-        return self.votes.where(Vote.user == user & Vote.is_upvote == False).exists()
+        return self.votes.where(Vote.user == user).where(Vote.is_upvote == False).exists()
     
     def has_user_voted(self, user: User):
         """Checks if a given user has voted."""
-        return self.has_user_upvoted(user) | self.has_user_downvoted(user)
+        return self.has_user_upvoted(user) or self.has_user_downvoted(user)
 
     def get_score(self):
         """Gets the sum of all the votes (upvotes - downvotes)."""
